Put default case into closure
In the original C version of match, there is a label after the default
case that can be jumped into from above.  The goto implementation in Go
does not allow us to jump to that label in all cases.  So instead, we
form a closure from the default case that we can call into anyplace.